Image Changes
In my section about the Pulic Theater I talked about the changes made to their logo over time and rather than just have my written explanation I thought it would be better to have a visual aid so I used an image that showed the changes over time. I used the picture tag to change the image back to a picture of their only their current logo when seen on mobile sizes (800px or less) as it became hard to read the image text.